+++ packages/flame/CHANGELOG.md
@@ -55,8 +55,7 @@
  - `ComponentEffect` is now a `Component`
  - `HasGameRef` can now operate independently from `Game`
  - `initialDelay` and `peakDelay` for effects to handle time before and after an effect
- - Removes `component.onMount` since `onLoad` can be used for all its purposes
- - Add `component.onParentChange` which runs after `component.prepare` and `component.onLoad`
+ - `component.onMount` now runs every time a component gets a new parent
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps